requirements.txt
Django>=2.2.2,<2.3.0
djangorestframework>=3.9.4,<3.10.0
flake8>=3.7.7,<3.8.0
psycopg2-binary>=2.8.2,<2.9.0
django-cors-headers<=3.0.2,<3.1.0
django-filter>=2.2.0,<2.3.0
django-simple-history>=2.7.3,<2.8.0
coverage<=4.5.3,<4.6.0
django-decouple==2.1
django-s3-storage==0.12.4
zappa_django_utils==0.4.0